fix replaygain, use floats for replaygain, fix navigation bar not showing current screen selection, add translations for playing and search screen

+++ app/src/main/java/com/craftworks/music/data/model/MediaData.kt
@@ -10,8 +10,8 @@ data class MusicFolder (
 
 @Serializable
 data class GainInfo(
-    val album: Double? = null,
-    val track: Double? = null
+    val album: Float? = null,
+    val track: Float? = null
 )
 
 data class PlaylistRules(
